graph is less specific
